Quote port "21:21" throughout, add note about YAML base-60 parsing, add compose example
When mapping ports in the HOST:CONTAINER format, you may experience erroneous results when using a container port lower than 60, because YAML parses numbers in the format xx:yy as a base-60 value. For this reason, it is recommended to always explicitly specify low-number port mappings as strings.

Port 21 is impacted by this, and quoting the port ensures that, should someone translate this to a docker-compose file, this behaviour won't arise.  As such, the 21:21 is quoted in the various examples in the readme file.

Moreover, I have translated the first usage example into a compose file and placed a note there as well.